+++ plugins/qeta-backend/src/service/routes/questions.ts
@@ -1,17 +1,20 @@
 import { RouterOptions } from '../router';
 import { QuestionsOptions } from '../../database/QetaStore';
 import {
-  checkPermissions,
+  authorize,
   getCreated,
   getUsername,
-  isModerator,
   mapAdditionalFields,
 } from '../util';
 import Ajv from 'ajv';
 import { Request, Router } from 'express';
 import {
+  qetaCreateCommentPermission,
   qetaCreateQuestionPermission,
-  qetaReadPermission,
+  qetaDeleteCommentPermission,
+  qetaDeleteQuestionPermission,
+  qetaEditQuestionPermission,
+  qetaReadQuestionPermission,
 } from '@drodil/backstage-plugin-qeta-common';
 import addFormats from 'ajv-formats';
 import {
@@ -31,7 +34,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   router.get(`/questions`, async (request, response) => {
     // Validation
     const username = await getUsername(request, options, true);
-    await checkPermissions(request, qetaReadPermission, options);
+    // TODO: Get Query from permission decision
     const validateQuery = ajv.compile(QuestionsQuerySchema);
     if (!validateQuery(request.query)) {
       response
@@ -51,7 +54,8 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   router.get(`/questions/list/:type`, async (request, response) => {
     // Validation
     const username = await getUsername(request, options, true);
-    await checkPermissions(request, qetaReadPermission, options);
+    await authorize(request, qetaReadQuestionPermission, options);
+    // TODO: Get Query from permission decision
     const validateQuery = ajv.compile(QuestionsQuerySchema);
     if (!validateQuery(request.query)) {
       response
@@ -88,8 +92,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
     // Validation
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    await checkPermissions(request, qetaReadPermission, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -97,6 +99,12 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const question = await database.getQuestion(
       username,
@@ -108,9 +116,11 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
-    question.answers?.map(a =>
-      mapAdditionalFields(username, a, options, moderator),
+    await mapAdditionalFields(request, question, options);
+    await Promise.all(
+      (question.answers ?? []).map(
+        async a => await mapAdditionalFields(request, a, options),
+      ),
     );
 
     signalQuestionStats(signals, question);
@@ -124,9 +134,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
     // Validation
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const created = await getCreated(request, options);
-    await checkPermissions(request, qetaReadPermission, options);
     const validateRequestBody = ajv.compile(CommentSchema);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
@@ -142,6 +150,14 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .send({ errors: validateRequestBody.errors, type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+    await authorize(request, qetaCreateCommentPermission, options);
+
     const question = await database.commentQuestion(
       questionId,
       username,
@@ -154,9 +170,11 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
-    question.answers?.map(a =>
-      mapAdditionalFields(username, a, options, moderator),
+    await mapAdditionalFields(request, question, options);
+    await Promise.all(
+      (question.answers ?? []).map(
+        async a => await mapAdditionalFields(request, a, options),
+      ),
     );
 
     if (events) {
@@ -182,8 +200,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       // Validation
       // Act
       const username = await getUsername(request, options);
-      const moderator = await isModerator(request, options);
-      await checkPermissions(request, qetaReadPermission, options);
+
       const questionId = Number.parseInt(request.params.id, 10);
       const commentId = Number.parseInt(request.params.commentId, 10);
       if (Number.isNaN(questionId) || Number.isNaN(commentId)) {
@@ -191,11 +208,24 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         return;
       }
 
+      await authorize(
+        request,
+        qetaDeleteQuestionPermission,
+        options,
+        request.params.id,
+      );
+
+      await authorize(
+        request,
+        qetaDeleteCommentPermission,
+        options,
+        request.params.commentId,
+      );
+
       const question = await database.deleteQuestionComment(
         questionId,
         commentId,
         username,
-        moderator,
       );
 
       if (question === null) {
@@ -203,9 +233,11 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         return;
       }
 
-      mapAdditionalFields(username, question, options, moderator);
-      question.answers?.map(a =>
-        mapAdditionalFields(username, a, options, moderator),
+      await mapAdditionalFields(request, question, options);
+      await Promise.all(
+        (question.answers ?? []).map(
+          async a => await mapAdditionalFields(request, a, options),
+        ),
       );
 
       // Response
@@ -242,7 +274,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   // POST /questions
   router.post(`/questions`, async (request, response) => {
     // Validation
-    await checkPermissions(request, qetaCreateQuestionPermission, options);
     const validateRequestBody = ajv.compile(PostQuestionSchema);
     if (!validateRequestBody(request.body)) {
       response
@@ -250,6 +281,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .json({ errors: validateRequestBody.errors, type: 'body' });
       return;
     }
+    await authorize(request, qetaCreateQuestionPermission, options);
 
     const tags = getTags(request);
     const entities = getEntities(request);
@@ -302,12 +334,16 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       return;
     }
 
+    await authorize(
+      request,
+      qetaEditQuestionPermission,
+      options,
+      request.params.id,
+    );
+
     const tags = getTags(request);
     const entities = getEntities(request);
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
-    const globalEdit =
-      options.config.getOptionalBoolean('qeta.allowGlobalEdits') ?? false;
 
     // Act
     const question = await database.updateQuestion(
@@ -318,7 +354,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       tags,
       entities,
       request.body.images,
-      moderator || globalEdit,
     );
 
     if (!question) {
@@ -345,7 +380,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   // DELETE /questions/:id
   router.delete('/questions/:id', async (request, response) => {
     // Validation
-    const moderator = await isModerator(request, options);
     const username = await getUsername(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
@@ -354,6 +388,12 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaDeleteQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     if (events) {
       const question = database.getQuestion(username, questionId, false);
@@ -368,11 +408,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
     }
 
     // Act
-    const deleted = await database.deleteQuestion(
-      username,
-      questionId,
-      moderator,
-    );
+    const deleted = await database.deleteQuestion(questionId);
 
     // Response
     response.sendStatus(deleted ? 200 : 404);
@@ -392,9 +428,15 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       return;
     }
 
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
+
     // Act
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const voted = await database.voteQuestion(username, questionId, score);
 
     if (!voted) {
@@ -409,7 +451,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       return;
     }
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
     question.ownVote = score;
 
     if (events) {
@@ -443,7 +485,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   // GET /questions/:id/favorite
   router.get(`/questions/:id/favorite`, async (request, response) => {
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -451,6 +492,12 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const favorited = await database.favoriteQuestion(username, questionId);
 
@@ -465,7 +512,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       false,
     );
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
 
     // Response
     response.json(question);
@@ -474,7 +521,6 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
   // GET /questions/:id/unfavorite
   router.get(`/questions/:id/unfavorite`, async (request, response) => {
     const username = await getUsername(request, options);
-    const moderator = await isModerator(request, options);
     const questionId = Number.parseInt(request.params.id, 10);
     if (Number.isNaN(questionId)) {
       response
@@ -482,6 +528,12 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
         .send({ errors: 'Invalid question id', type: 'body' });
       return;
     }
+    await authorize(
+      request,
+      qetaReadQuestionPermission,
+      options,
+      request.params.id,
+    );
 
     const unfavorited = await database.unfavoriteQuestion(username, questionId);
 
@@ -496,7 +548,7 @@ export const questionsRoutes = (router: Router, options: RouterOptions) => {
       false,
     );
 
-    mapAdditionalFields(username, question, options, moderator);
+    await mapAdditionalFields(request, question, options);
 
     // Response
     response.json(question);
